pyproject.toml
[tool.poetry]
name = "ftt"
version = "0.1.0-alpha+001"
description = "Financial Trading Tool (FTT) – is an asset management application that helps to make the right decision on time."
authors = ["Artem Melnykov <melnykov.artem.v@gmail.com>"]
classifiers = [
    "Topic :: Office/Business :: Financial :: Investment",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Financial and Insurance Industry",
]
packages = [
    { include = "ftt" },
]

include = [
    "LICENSE.md",
]

readme = "README.md"

[tool.poetry.urls]
"Homepage" = "https://github.com/ftt-project/ftt"
"Bug Tracker" = "https://github.com/ftt-project/ftt/issues"


[tool.poetry.dependencies]
python = "^3.9"
yfinance = "^0.1.54"
pandas = "^1.4.0"
pyyaml = "^6.0"
pandas_datareader = "^0.9"
Yahoo-ticker-downloader = "^3.0.1"
peewee = "^3.14.0"
rich = "^11.1"
pendulum = "^2.1.2"
cvxopt = "1.2.7"
pyportfolioopt = "^1.2.7"
python-nubia = "^0.2b5"
result = "^0.7.0"
Riskfolio-Lib = "^2.0.0"
SQLAlchemy = "^1.4.31"
pyinstrument = "^4.1.1"
ibapi = "^9.81.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
isort = "^5.9.1"
pytest-mock = "^3.5.1"
pytest = "^6.2.3"
flake8 = "^4.0.0"
black = "^22.1.0"
pylint = "^2.6.0"
pytest-cov = "^3.0.0"
mypy = "^0.931"

[tool.poetry.scripts]
ftt = "ftt:__main__"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
addopts="""
--cov-report term-missing \
--cov ftt \
--cov tests \
--ignore tests/testcommon.py \
--ignore tests/helper.py \
--no-cov-on-fail \
-ra \
-W ignore"""